Theme controller finished
