--- src/lib/stores/theme.js
@@ -1,3 +1,0 @@
-import { writable } from 'svelte/store';
-
-export const lightMode = writable(true);
