+++ src/routes/+layout.svelte
@@ -1,5 +1,16 @@
 <script lang="ts">
     import '../app.css';
+    import { lightMode, toggle } from '$lib/stores/theme';
+    import { onMount } from 'svelte';
+    import { page } from '$app/stores';
+
+    onMount(() => ($lightMode = localStorage.getItem("theme") === "true"));
 </script>
 
-<slot />
+{#if $page.url.pathname !== '/'}
+    <div data-theme={$lightMode ? 'light' : 'dark'}>
+        <slot />
+    </div>
+{:else}
+    <slot />
+{/if}
