+++ src/routes/+page.svelte
@@ -9,8 +9,8 @@
     <p class="flex justify-center text-xl">Sua plataforma de arquitetura</p>
 </div>
 <div class="flex justify-center gap-4">
-    <a href="/auth/login/" class="btn btn-primary w-36">Iniciar Sessão</a>
-    <a href="/auth/register/" class="btn btn-neutral w-36 text-primary">Criar Conta</a>
+    <a href="/login" class="btn btn-primary w-36">Iniciar Sessão</a>
+    <a href="/register" class="btn btn-neutral w-36 text-primary">Criar Conta</a>
 </div>
 <div class="flex justify-center">
     <img src="/gallery/plano.png" alt="Planta" />
